requirements.txt
pypyodbc
asn1crypto==1.5.1
backoff==1.10.0
bcrypt==4.0.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
ciso8601==2.3.0
cryptography==41.0.3
dataconf==2.2.1
dateparser==1.1.8
Deprecated==1.2.14
filelock==3.12.3
flake8==6.1.0
freezegun==0.3.15
idna==3.4
keboola.component==1.4.3
keboola.csvwriter==1.0.1
mccabe==0.7.0
mock==4.0.2
mysql-replication==0.43.0
oscrypto==1.3.0
packaging==23.1
paramiko==3.3.1
pendulum==1.2.0
platformdirs==3.8.1
pycodestyle==2.11.0
pycparser==2.21
pycryptodomex==3.18.0
pyflakes==3.1.0
pygelf==0.4.2
pyhocon==0.3.60
PyJWT==2.8.0
PyMySQL==1.1.0
PyNaCl==1.5.0
pyOpenSSL==23.2.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3
pytzdata==2020.1
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
simplejson==3.19.1
six==1.16.0
sortedcontainers==2.4.0
sqlparse==0.4.4
sshtunnel==0.4.0
tomlkit==0.12.1
typing-extensions==4.7.1
tzlocal==5.0.1
urllib3==1.26.16
wrapt==1.15.0
https://bitbucket.org/kds_consulting_team/datadirtest/get/1.7.1.zip#egg=datadirtest
